Handle multiple items with same search text
When ISelectionAdapter.Commit was completed the code used
UpdateTextCompletion to find the first item with a matching string. If
the items source contained multiple items with a matching string, you
only ever got the first item, regardless of which one you selected in
the SelectionAdapter.
This change uses the specific item selected in the SelectionAdapter
instead of trying to "find" a matching item.